# src/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HarborGuard API",
    "version": "0.1b",
    "description": "HarborGuard Container Security Platform API Documentation",
    "contact": {
      "name": "HarborGuard Team",
      "url": "https://github.com/HarborGuard/HarborGuard"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/HarborGuard/HarborGuard/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/scans/start": {
      "post": {
        "summary": "Start a new container scan",
        "description": "Initiates a security scan for a container image",
        "tags": [
          "Scans"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "image",
                  "tag"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "Image name",
                    "example": "nginx"
                  },
                  "tag": {
                    "type": "string",
                    "description": "Image tag",
                    "example": "latest"
                  },
                  "source": {
                    "type": "string",
                    "enum": [
                      "registry",
                      "local"
                    ],
                    "description": "Image source"
                  },
                  "dockerImageId": {
                    "type": "string",
                    "description": "Docker image ID for local images"
                  },
                  "repositoryId": {
                    "type": "string",
                    "description": "Repository ID for private registries"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "requestId": {
                      "type": "string",
                      "description": "Unique scan request ID"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/scans/upload": {
      "post": {
        "summary": "POST /api/scans/upload",
        "tags": [
          "Scans"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "type": "string",
                    "description": "requestId parameter"
                  },
                  "image": {
                    "type": "object",
                    "description": "image parameter"
                  },
                  "tag": {
                    "type": "string",
                    "description": "tag parameter"
                  },
                  "registry": {
                    "type": "string",
                    "description": "registry parameter (optional)"
                  },
                  "digest": {
                    "type": "string",
                    "description": "digest parameter"
                  },
                  "platform": {
                    "type": "string",
                    "description": "platform parameter (optional)"
                  },
                  "sizeBytes": {
                    "type": "number",
                    "description": "sizeBytes parameter (optional)"
                  },
                  "scan": {
                    "type": "object",
                    "description": "scan parameter"
                  },
                  "finishedAt": {
                    "type": "string",
                    "description": "finishedAt parameter (optional)"
                  },
                  "status": {
                    "type": "string",
                    "description": "status parameter"
                  },
                  "reportsDir": {
                    "type": "string",
                    "description": "reportsDir parameter (optional)"
                  },
                  "errorMessage": {
                    "type": "string",
                    "description": "errorMessage parameter (optional)"
                  },
                  "scannerVersions": {
                    "type": "string",
                    "description": "scannerVersions parameter"
                  },
                  "scanConfig": {
                    "type": "string",
                    "description": "scanConfig parameter"
                  },
                  "reports": {
                    "type": "object",
                    "description": "reports parameter (optional)"
                  },
                  "grype": {
                    "type": "string",
                    "description": "grype parameter (optional)"
                  },
                  "syft": {
                    "type": "string",
                    "description": "syft parameter (optional)"
                  },
                  "dockle": {
                    "type": "string",
                    "description": "dockle parameter (optional)"
                  },
                  "metadata": {
                    "type": "string",
                    "description": "metadata parameter (optional)"
                  }
                },
                "required": [
                  "requestId",
                  "image",
                  "tag",
                  "digest",
                  "scan",
                  "status",
                  "scannerVersions",
                  "scanConfig"
                ]
              }
            }
          }
        }
      },
      "get": {
        "summary": "GET /api/scans/upload",
        "tags": [
          "Scans"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
  },
...
}